Added tests for eventfilereader.py
